requirements.txt
pyCryptodome
requests
PyQt5==5.11.3